+++ light/LED.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2021 The LineageOS Project
+ * Copyright (C) 2021-2022 The LineageOS Project
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -18,8 +18,8 @@ public:
     LED(std::string type);
 
     bool exists();
-    bool setBreath(uint32_t value);
-    bool setBrightness(uint32_t value);
+    bool setBreath(uint8_t value);
+    bool setBrightness(uint8_t value);
 private:
     std::string mBasePath;
     uint32_t mMaxBrightness;
