+++ light/Utils.h
@@ -1,5 +1,5 @@
 /*
- * Copyright (C) 2021 The LineageOS Project
+ * Copyright (C) 2021-2022 The LineageOS Project
  *
  * SPDX-License-Identifier: Apache-2.0
  */
@@ -14,10 +14,10 @@ namespace hardware {
 namespace light {
 
 typedef struct argb {
-    uint32_t alpha;
-    uint32_t red;
-    uint32_t green;
-    uint32_t blue;
+    uint8_t alpha;
+    uint8_t red;
+    uint8_t green;
+    uint8_t blue;
 } argb_t;
 
 bool fileWriteable(const std::string& file);
@@ -26,8 +26,8 @@ bool readFromFile(const std::string& file, uint32_t *content);
 bool writeToFile(const std::string& file, uint32_t content);
 bool isLit(uint32_t color);
 argb_t colorToArgb(uint32_t color);
-uint32_t argbToBrightness(argb_t c_argb);
-uint32_t colorToBrightness(uint32_t color);
+uint8_t argbToBrightness(argb_t c_argb);
+uint8_t colorToBrightness(uint32_t color);
 
 } // namespace light
 } // namespace hardware
